server/segments_endpoint.go
package main

import (
	"database/sql"
	"encoding/json"
	"net/http"

	"github.com/dadikovi/race-timer/server/core"
)

type SegmentEndpoint struct{}

func (se *SegmentEndpoint) create(w http.ResponseWriter, r *http.Request, db *sql.DB) {
	var s core.Segment

	decoder := json.NewDecoder(r.Body)
	if err := decoder.Decode(&s); err != nil {
		respondWithError(w, http.StatusBadRequest, "Invalid request payload")
		return
	}
	defer r.Body.Close()

	if err := s.Save(db); err != nil {
		respondWithError(w, http.StatusInternalServerError, err.Error())
		return
	}

	respondWithJSON(w, http.StatusCreated, s)
}

func respondWithError(w http.ResponseWriter, code int, message string) {
	respondWithJSON(w, code, map[string]string{"error": message})
}

func respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
	response, _ := json.Marshal(payload)

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)
	w.Write(response)
}
